% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobuf_filter.R
\name{addProtobufFilter}
\alias{addProtobufFilter}
\title{Add a Protocol Buffer filter.}
\usage{
addProtobufFilter(api, name = "protoBuf", descriptor_path)
}
\arguments{
\item{api}{A plumber api.}

\item{name}{What to call the filter. Default: "protoBuf".}

\item{descriptor_path}{Path to the .proto descriptor file.}
}
\description{
Adds a filter for unserializing Protocol Buffer messages
    to your plumber API. Requires a .proto descriptor file.
}
