% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobuf_serializer.R
\name{addProtobufSerializer}
\alias{addProtobufSerializer}
\title{Add a Protocol Buffer serializer.}
\usage{
addProtobufSerializer(name = "protoBuf",
  serializer = protobuf_serializer)
}
\arguments{
\item{name}{The name of the serializer (character string). Default: "ProtoBuf"}

\item{serializer}{The serializer to be added. Default: protobuf_serializer}
}
\description{
This is a convenience wrapper for plumber::addSerializer.
}
