% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobuf_serializer.R
\name{protobuf_serializer}
\alias{protobuf_serializer}
\title{Protocol Buffer serializer.}
\usage{
protobuf_serializer()
}
\description{
Serializes Protocol Buffer messages, and adds the message type
    to the response header.
}
